Fixing go vet errors
